requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==23.2.0
autopep8==2.1.0
blinker==1.8.2
CacheControl==0.14.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
dnspython==2.4.2
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gcloud==0.18.3
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-cloud-vision==3.7.4
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jwcrypto==1.5.6
MarkupSafe==3.0.2
msgpack==1.1.0
msgspec==0.18.6
multidict==6.0.5
numpy==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.6.1
pyparsing==3.2.0
Pyrebase4==4.8.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
selenium==4.20.0
setuptools==75.2.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
tqdm==4.66.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.16.0
